pyproject.toml
[tool.poetry]
name = "kbpartpicker-backend"
version = "0.1.0"
description = ""
authors = ["vimiomori <vivian.muchen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
attr = "^0.3.1"
attrs = "^20.3.0"
Flask-Admin = "^1.5.7"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^2.6.0"
inflect = "^5.0.2"
lxml = "^4.6.2"
Mako = "^1.1.4"
psycopg2-binary = "^2.8.6"
python-dateutil = "^2.8.1"
python-editor = "^1.0.4"
selenium = "^3.141.0"
urllib3 = "^1.26.3"
WTForms = "^2.3.3"
sanic = "^20.12.1"
Flask-API = "^2.0"
Flask-RESTful = "^0.3.8"
inflection = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest-sanic = "^1.6.2"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
